observability/test-app/requirements.txt
fastapi
uvicorn
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-distro
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
prometheus_client